requirements.txt
Django==3.2.10
python-dateutil
tzlocal